+++ internal_filesystem/lib/mpos/board/qemu.py
@@ -122,11 +122,10 @@ def keypad_read_cb(indev, data):
     # While in an on-screen keyboard, PREV button is LEFT and NEXT button is RIGHT
     focus_group = lv.group_get_default()
     focus_keyboard = False
-    if focus_group:
-        current_focused = focus_group.get_focused()
-        if isinstance(current_focused, lv.keyboard):
-            #print("focus is on a keyboard")
-            focus_keyboard = True
+    current_focused = focus_group.get_focused()
+    if isinstance(current_focused, lv.keyboard):
+        #print("focus is on a keyboard")
+        focus_keyboard = True
 
     if near_simul:
         current_key = lv.KEY.ENTER
@@ -195,8 +194,7 @@ def keypad_read_cb(indev, data):
     if data.state == lv.INDEV_STATE.PRESSED and data.key == lv.KEY.ESC:
         mpos.ui.back_screen()
 
-group = lv.group_create()
-group.set_default()
+group = lv.group_get_default()
 
 # Create and set up the input device
 indev = lv.indev_create()